feat: divided VciBaseEndpoint and BaseEndpoint
