+++ pyeudiw/openid4vci/endpoints/notification_endpoint.py
@@ -4,7 +4,7 @@ from satosa.response import (
 )
 
 from pyeudiw.jwt.jws_helper import JWSHelper
-from pyeudiw.openid4vci.endpoints.base_endpoint import BaseEndpoint
+from pyeudiw.openid4vci.endpoints.vci_base_endpoint import VCIBaseEndpoint
 from pyeudiw.openid4vci.models.notification_request import NotificationRequest
 from pyeudiw.openid4vci.tools.exceptions import (
     InvalidRequestException,
@@ -20,7 +20,7 @@ from pyeudiw.tools.content_type import (
 )
 
 
-class NotificationHandler(BaseEndpoint):
+class NotificationHandler(VCIBaseEndpoint):
 
     def __init__(self, config: dict, internal_attributes: dict[str, dict[str, str | list[str]]], base_url: str, name: str):
         """
